Fix a bug of the color and font panels positioning, when the page have been scrolled.
The jQuery "offset()" method return the position of an element relative to the document, not the browser window. Although the "left" and "top" css properties of a "position: fixed;" element, set the position relative to the browser window. I have added the scroll variation to the offset, with the "scrollTop()" and "scrollLeft()" jQuery methods.